gw: add wgs84 distance

+++ backend/src/main/kotlin/ru/dmitry4k/geomarkback/service/Distance.kt
@@ -3,5 +3,5 @@ package ru.dmitry4k.geomarkback.service
 import ru.dmitry4k.geomarkback.dto.GeoPoint
 
 interface Distance {
-    fun distance(a: GeoPoint, b: GeoPoint): Long
+    fun distance(a: GeoPoint, b: GeoPoint): Double
 }
